pyproject.toml
[tool.poetry]
name = "faststream-template"
version = "0.1.0"
description = ""
authors = ["Selix <asel1x@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.114.0"}
sqlmodel = "^0.0.22"
faststream = {extras = ["rabbit"], version = "^0.5.14"}
alembic = "^1.13.2"
pyjwt = "^2.8.0"
dishka = "^1.2.0"
greenlet = "^3.0.3"
pydantic-settings = "^2.3.4"
loguru = "^0.7.2"
bcrypt = "^4.1.3"
asyncpg = "^0.29.0"
pre-commit = "^3.7.1"
taskiq = "^0.11.6"
taskiq-faststream = "^0.2.0"
redis = "^5.0.8"
types-redis = "^4.6.0.20240819"
pytest-asyncio = "^0.24.0"
aiosqlite = "^0.20.0"
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
